Add files via upload
The aim of this project is to develop a system that can perform the prediction of diabetes for a patient by combining the results of different machine learning algorithms. The algorithms like K Nearest Neighbour, SVC, Logistic Regression, Decision Tree, AdaBoost, Random Forest, Naive Bayes, and XGBClassifier are used. The accuracy of the model using each of the algorithms is calculated.